Refactor the playbook options into a mixin module
