+++ config/routes.rb
@@ -284,6 +284,7 @@ Rails.application.routes.draw do
         atomic_form_field_changed
         atomic_st_edit
         automate_button_field_changed
+        playbook_options_field_changed
         explorer
         get_ae_tree_edit_key
         group_create
@@ -2202,6 +2203,7 @@ Rails.application.routes.draw do
         ae_tree_select_toggle
         accordion_select
         automate_button_field_changed
+        playbook_options_field_changed
         cancel_import
         change_tab
         dialog_edit
